Dedicated time utilities module
