+++ src/args.rs
@@ -7,6 +7,8 @@ use pnet::datalink::MacAddr;
 use pnet::packet::arp::{ArpHardwareType, ArpOperation};
 use pnet::packet::ethernet::EtherType;
 
+use crate::time::parse_to_milliseconds;
+
 const TIMEOUT_MS_FAST: u64 = 800;
 const TIMEOUT_MS_DEFAULT: u64 = 2000;
 
@@ -419,102 +421,6 @@ impl ScanOptions {
     pub fn has_vlan(&self) -> bool {
 
         matches!(&self.vlan_id, Some(_)) 
-    }
-
-}
-
-/**
- * Parse a given time string into milliseconds. This can be used to convert a
- * string such as '20ms', '10s' or '1h' into adequate milliseconds. Without
- * suffix, the default behavior is to parse into milliseconds.
- */
-fn parse_to_milliseconds(time_arg: &str) -> Result<u64, &str> {
-
-    let len = time_arg.len();
-
-    if time_arg.ends_with("ms") {
-        let milliseconds_text = &time_arg[0..len-2];
-        return match milliseconds_text.parse::<u64>() {
-            Ok(ms_value) => Ok(ms_value),
-            Err(_) => Err("invalid milliseconds")
-        };
-    }
-
-    if time_arg.ends_with('s') {
-        let seconds_text = &time_arg[0..len-1];
-        return match seconds_text.parse::<u64>().map(|value| value * 1000) {
-            Ok(ms_value) => Ok(ms_value),
-            Err(_) => Err("invalid seconds")
-        };
-    }
-
-    if time_arg.ends_with('m') {
-        let seconds_text = &time_arg[0..len-1];
-        return match seconds_text.parse::<u64>().map(|value| value * 1000 * 60) {
-            Ok(ms_value) => Ok(ms_value),
-            Err(_) => Err("invalid minutes")
-        };
-    }
-
-    if time_arg.ends_with('h') {
-        let hour_text = &time_arg[0..len-1];
-        return match hour_text.parse::<u64>().map(|value| value * 1000 * 60 * 60) {
-            Ok(ms_value) => Ok(ms_value),
-            Err(_) => Err("invalid hours")
-        };
-    }
-
-    match time_arg.parse::<u64>() {
-        Ok(ms_value) => Ok(ms_value),
-        Err(_) => Err("invalid milliseconds")
-    }
-}
-
-#[cfg(test)]
-mod tests {
-
-    use super::*;
-
-    #[test]
-    fn should_parse_milliseconds() {
-        
-        assert_eq!(parse_to_milliseconds("1000"), Ok(1000));
-    }
-
-    #[test]
-    fn should_parse_seconds() {
-        
-        assert_eq!(parse_to_milliseconds("5s"), Ok(5000));
-    }
-
-    #[test]
-    fn should_parse_minutes() {
-        
-        assert_eq!(parse_to_milliseconds("3m"), Ok(180_000));
-    }
-
-    #[test]
-    fn should_parse_hours() {
-        
-        assert_eq!(parse_to_milliseconds("2h"), Ok(7_200_000));
-    }
-
-    #[test]
-    fn should_deny_negative() {
-        
-        assert_eq!(parse_to_milliseconds("-45"), Err("invalid milliseconds"));
-    }
-
-    #[test]
-    fn should_deny_floating_numbers() {
-        
-        assert_eq!(parse_to_milliseconds("3.235"), Err("invalid milliseconds"));
-    }
-
-    #[test]
-    fn should_deny_invalid_characters() {
-        
-        assert_eq!(parse_to_milliseconds("3z"), Err("invalid milliseconds"));
     }
 
 }
